aarch64: Add USB storage devices
RH-Author: Andrew Jones <drjones@redhat.com>
Message-id: <20210803114113.19126-2-drjones@redhat.com>
Patchwork-id: 101950
O-Subject: [RHEL-AV-8.5.0 qemu-kvm PATCH v2 1/1] aarch64: Add USB storage devices
Bugzilla: 1974579
RH-Acked-by: Gavin Shan <gshan@redhat.com>
RH-Acked-by: Philippe Mathieu-Daudé <philmd@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

There's no reason not to support these devices and doing so allows us
to match x86 and ppc. Also to match the other architectures we do not
enable CONFIG_USB_STORAGE_BOT, as that was disabled for the other
architectures for BZ1866133.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>