redhat: Adding slirp to the exploded tree
RH-Author: Danilo de Paula <ddepaula@redhat.com>
Message-id: <20190907020756.8619-1-ddepaula@redhat.com>
Patchwork-id: 90309
O-Subject: [RHEL-AV-8.1.0 qemu-kvm PATCH 1/1] redhat: Adding slirp to the exploded tree
Bugzilla:
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

Until qemu-kvm-3.1 slirp used to live as a regular folder in qemu-kvm.
After that it got moved into its own submodule. Which means it's not
part of the qemu-kvm git tree anymore.

This passed unoticed for RHEL-AV-8.0.1 and 8.1.0 because qemu still ships
the code in the tarball. That's why scratch builds still works (it's based in
the tarball content).

As we're receiving some CVE's against slirp, we need a way to patch
slirp in RHEL-8.1.0 without handling as a separate package (as we do for
firmwares).

The simplest solution is to copy the slirp folder from the tarball into the
exploded tree.

To be able to do that, I had to make some changes:

slirp needs to be removed from .gitmodules, otherwise git complains
about files on it.

Since "make -C redhat rh-brew" uses the tarball and apply all the
patches on top of it, we need to remove the folder from the tarball before applying
the patch (because we are actually re-applying them).

We also need to use --ignore-submodule while generating the patches for
scratch-build, otherwise it will include some weird definition of the
slirp folder in the patch, something that /usr/bin/patch gets mad with.

After that I compared the patch list, after and before this change, and
saw no major differences.

This is an exploded-tree-only change and shouldn't be applied to dist-git.

Signed-off-by: Danilo C. L. de Paula <ddepaula@redhat.com>

Rebase notes (weekly-210217):
 - Upstream slirp updated to 8f43a99191afb47ca3f3c6972f6306209f367ece

Rebase notes (6.1.0-rc2):
- Upstream slirp updated to a88d9ace234a24ce1c17189642ef9104799425e0

Merged commits (weekly-210203):
 - a3f5f082f Drop bogus IPv6 messagesa

Merged commits (weekly-210714):
- ce9ddeef04 Add mtod_check()
- 0609398e76 bootp: limit vendor-specific area to input packet memory buffer
- 377f755273 bootp: check bootp_input buffer size
- 4101e41f0d upd6: check udp6_input buffer size
- 7a663c9667 tftp: check tftp_input buffer size
- 76f81fc22c tftp: introduce a header structure
- 6903e9ba25 udp: check upd_input buffer size
- 8aa4fe0b6d Fix "DHCP broken in libslirp v4.6.0"