spec: Remove block-curl and block-ssh dependency
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 36: Remove block-curl and block-ssh dependency
RH-Commit: [1/1] 50afa433e183dc9e61028e2ef55c2974d5814490 (mrezanin/rhel-src-qemu-kvm)
RH-Bugzilla: 2001100

block-curl and block-ssh were added to support virt-v2v. However,
virt-v2v switched to nbd plugins instead in RHEL-AV 8.3.0 that are
more featureful.

We don't plan to completely remove support for these modules but
we are moving them to CRB. To be able to this, we have to remove
dependency to subpackages providing these modules.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>