tools/virtiofsd: Add fstatfs64 syscall to the seccomp allowlist
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 44: Apply RHEL 9.0.0 Beta fixes to RHEL 9.0.0
RH-Commit: [2/2] 0085289cefb57d49d2423b4f3376e8cf4a970012 (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 2005026
RH-Acked-by: Daniel P. Berrangé <berrange@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

The virtiofsd currently crashes on s390x when doing something like
this in the guest:

 mkdir -p /mnt/myfs
 mount -t virtiofs myfs /mnt/myfs
 touch /mnt/myfs/foo.txt
 stat -f /mnt/myfs/foo.txt

The problem is that the fstatfs64 syscall is called in this case
from the virtiofsd. We have to put it on the seccomp allowlist to
avoid that the daemon gets killed in this case.

(cherry picked from commit 8cfd339b3d402f913fe520a4f35f30152fb4fb80)
Suggested-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>

patch_name: kvm-tools-virtiofsd-Add-fstatfs64-syscall-to-the-seccomp.patch
present_in_specfile: true
location_in_specfile: 21