spec: Remove qemu-virtiofsd
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 76: spec: Remove qemu-virtiofsd
RH-Commit: [1/1] 21be974f7632e1cf5640bc7887b05edfaf6d91ce (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 2055284
RH-Acked-by: Sergio Lopez <slp@redhat.com>
RH-Acked-by: Jon Maloy <jmaloy@redhat.com>

We are now providing virtiofsd as separate components that obsoletes
qemu version. As gating checks verify all built packages are usable
they fail on qemu-virtiofsd that can't be installed due to presence
of standalone version.

Removing build of the package to allow tests to pass.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>