spec: Obsolete ssh driver
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 120: Fixing upgrade path from RHEL 8
RH-Commit: [3/3] 179ad6f1807d8ef192e8baddc403c4f781e3c4b3 (mrezanin/rhel-src-qemu-kvm)
RH-Bugzilla: 2059175 2059146
RH-Acked-by: Daniel P. Berrangé <berrange@redhat.com>
RH-Acked-by: Jon Maloy <jmaloy@redhat.com>
RH-Acked-by: Eduardo Lima (Red Hat) <elima@rednat.com>

We removed dependency to ssh block driver in the RHEL 9 beta. However,
due to late handling package is still build and driver whitelisted.
We have only removed dependency to not install the package and be able
to remove it from compose.

However, this removal did not include obsoletes tag so upgrade path was
broken as the package is not properly removed when update is run and
ssh block driver is installed (as it require maching qemu-kvm-common
package).

Adding obsoletes tag to be able properly update.

Tracked by BZ 2059146.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>