spec: Obsolete ssh driver
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 78: Synchronize with RHEL 9.0.0 build qemu-kvm-6.2.0-11.el9_0.1
RH-Commit: [5/5] 69c58e0e3d12a0d26afc18c78972d95f3525ed58 (mrezanin/centos-src-qemu-kvm)
RH-Bugzilla: 2062828 2062819 2062817 2062813
RH-Acked-by: Stefano Garzarella <sgarzare@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

We removed dependency to ssh block driver in the RHEL 9 beta. However,
due to late handling package is still build and driver whitelisted.
We have only removed dependency to not install the package and be able
to remove it from compose.

However, this removal did not include obsoletes tag so upgrade path was
broken as the package is not properly removed when update is run and
ssh block driver is installed (as it require maching qemu-kvm-common
package).

Adding obsoletes tag to be able properly update.

Tracked by BZ 2059146.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>

Forward-port of RHEL 9.0.0 MR 120 (Fixing upgrade path from RHEL 8)