Set permission on installing files
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
RH-MergeRequest: 150: Set permission on installing files
RH-Commit: [1/1] 983fa0336090584374cf6ddb6cb8c0fe673931a8
RH-Bugzilla: 2072377
RH-Acked-by: Jon Maloy <jmaloy@redhat.com>
RH-Acked-by: Camilla Conte <None>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

When installing keymap files we do not set proper file permissions that
causes a lot of warnings to be printed out later as execution permission is
set. Setting proper permissions on installing to prevent these warnings.

In addition, test/Makefile.include is (do not know why) installed with
execution permission set. This cause warning on rpm build too. Changing
install command to not set the executable permissions.

All changes in this commit do not change anyting in generated rpms as
execution permission is stripped when mangling shebangs.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>